refactor: 重複・死にコードの HeaderContext / useI18nHeader を削除
## 背景
リポジトリ内に HeaderContext が 2 定義存在していた:
  - src/contexts/HeaderContext.ts
  - src/components/providers/HeaderProvider.tsx
前者は HeaderContext 本体と useHeader() を export しているものの、
どこからも参照されておらず、実質「HeaderConfig 型の置き場」として
のみ利用されていた。

また src/hooks/useI18nHeader.ts は前者から useHeaderConfig を
import していたが、それは export されていないため TypeScript
エラー状態で、かつ useI18nHeader 自体もどこからも呼ばれていない
完全な死コードだった。

## 変更
- 削除:
  - src/contexts/HeaderContext.ts
  - src/hooks/useI18nHeader.ts
- HeaderConfig 型 import を @/components/providers/HeaderProvider
  に差し替え (reservation 3 ページ + TransferInputStep):
  HeaderProvider 側の HeaderConfig はスーパーセットのため型的に安全。
- .storybook/decorators/withHeader.ts を実物の HeaderProvider 1 回
  ラップに簡素化 (2 系統 Context 両対応が不要になったため)。
- .storybook/README.md から 2 系統 Context に関する記述を削除。

## 補足
state 管理を行うのは HeaderProvider 側のみで、もう片方は同名の
型を持つだけだったので挙動は変わらない。
